requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.5 